requirements.txt
django==5.2.3
sqlparse
pymysql
djangorestframework
django-cors-headers
requests
duckdb
pyyaml
minio
pandas
pillow